requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
google-cloud-vision==3.5.0
python-multipart==0.0.6
Pillow==10.2.0
httpx==0.26.0
redis==5.0.1
python-jose[cryptography]==3.3.0
# New dependencies for Chalkie-inspired deck system
python-pptx==0.6.21
qdrant-client==1.7.0
sentence-transformers==2.2.2
requests==2.31.0
